Created layout.html
Created the layout file that will be inherited by all the other html files through jinja. Also added the jinja inheritance implementation in each of the added files. Still need to add it to book.html. Also tried to implement the style.css file into layout.html, however it was having errors and this has to be fixed at some point